feat(cookies): extract `CookieStore` from `NextResponse`
