+++ .changeset/rotten-moose-carry.md
@@ -1,0 +1,5 @@
+---
+'@edge-runtime/cookies': major
+---
+
+Extract `CookieStore` to its own class
