+++ packages/cookies/src/index.ts
@@ -1,3 +1,3 @@
-export type { Cookie, CookieListItem } from './serialize'
+export * from './cookie-store'
 export { ResponseCookies } from './response-cookies'
 export { RequestCookies } from './request-cookies'
